refactor: Structure refctoring WIP
